Move cryptokey validation to separate service part
